[MI-3828] Add feature to show linked channels and search the linked channels
