+++ webapp/src/components/Icon/Icon.map.tsx
@@ -433,4 +433,19 @@ export const IconMap : Record<IconName, JSX.Element> = {
             </defs>
         </svg>
     ),
+    lock: (
+        <svg
+            width='24'
+            height='24'
+            viewBox='0 0 24 24'
+            fill='none'
+            xmlns='http://www.w3.org/2000/svg'
+        >
+            <path
+                d='M16.5 17.9938V11.2438H7.5V17.9938H16.5ZM16.5 9.7498C16.92 9.7498 17.274 9.8938 17.562 10.1818C17.85 10.4698 17.994 10.8238 17.994 11.2438V17.9938C17.994 18.4138 17.844 18.7678 17.544 19.0558C17.256 19.3558 16.908 19.5058 16.5 19.5058H7.5C7.092 19.5058 6.738 19.3558 6.438 19.0558C6.15 18.7678 6.006 18.4138 6.006 17.9938V11.2438C6.006 10.8238 6.15 10.4698 6.438 10.1818C6.738 9.8938 7.092 9.7498 7.5 9.7498H8.256V8.2558C8.256 7.5598 8.418 6.9298 8.742 6.3658C9.078 5.7898 9.528 5.3338 10.092 4.9978C10.668 4.6618 11.304 4.4938 12 4.4938C12.696 4.4938 13.326 4.6618 13.89 4.9978C14.466 5.3338 14.916 5.7898 15.24 6.3658C15.576 6.9298 15.744 7.5598 15.744 8.2558V9.7498H16.5ZM12 6.0058C11.58 6.0058 11.196 6.1078 10.848 6.3118C10.512 6.5038 10.242 6.7738 10.038 7.1218C9.846 7.4578 9.75 7.8358 9.75 8.2558V9.7498H14.25V8.2558C14.25 7.8358 14.148 7.4578 13.944 7.1218C13.752 6.7738 13.482 6.5038 13.134 6.3118C12.798 6.1078 12.42 6.0058 12 6.0058Z'
+                fill='#3F4350'
+                fillOpacity='0.56'
+            />
+        </svg>
+    ),
 };
