+++ webapp/src/components/LinkedChannelCard/LinkedChannelCard.styles.scss
@@ -26,5 +26,17 @@
         width: 10px;
       }
     }
+
+    &__body {
+      width: 95%;
+    }
+
+    &__body-values {
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      max-width: 45%;
+      line-height: inherit;
+    }  
   }
 }
