+++ webapp/src/constants/common.constants.ts
@@ -8,6 +8,8 @@ export const defaultPage = 0;
 
 export const defaultPerPage = 20;
 
+export const debounceFunctionTimeLimit = 500;
+
 // Severity used in alert component
 export const alertSeverity: Record<SnackbarColor, SnackbarColor> = {
     success: 'success',
