--- webapp/src/containers/Rhs/views/ConnectAccount/index.ts
@@ -1,1 +1,0 @@
-export {ConnectAccount} from './ConnectAccount.container';
