--- webapp/src/containers/Rhs/views/ConnectedAccount/index.ts
@@ -1,1 +1,0 @@
-export {ConnectedAccount} from './ConnectedAccount.container';
