--- webapp/src/containers/Rhs/views/LinkedChannels/LinkedChannels.mock.ts
@@ -1,219 +1,0 @@
-export const mockLinkedChannels = [
-
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    }, {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-    {
-        msTeamsTeamID: 'string',
-        msTeamsTeamName: 'string',
-        msTeamsChannelID: 'string',
-        msTeamsChannelName: 'string',
-        mattermostTeamID: 'string',
-        mattermostTeamName: 'string',
-        mattermostChannelID: 'string',
-        mattermostChannelName: 'string',
-        mattermostChannelType: 'string',
-        msTeamsChannelType: 'string',
-    },
-
-];
