--- webapp/src/containers/Rhs/views/LinkedChannels/index.ts
@@ -1,1 +1,0 @@
-export {LinkedChannels} from './LinkedChannels.container';
