+++ webapp/src/styles/_utils.scss
@@ -104,6 +104,14 @@
     margin-top: 0;
 }
 
+.pt-0 {
+    padding-top: 0 !important;
+}
+
+.mt-10 {
+    margin-top: 10px;
+}
+
 .align-items-center {
     align-items: center;
 }
